requirements.txt
fastapi>=0.115,<1.0
python-dotenv>=1.0,<2.0
uvicorn[standard]>=0.34,<1.0
pymongo>=4.10,<5.0
dnspython>=2.7,<3.0
jinja2>=3.1,<4.0
python-multipart>=0.0.20,<1.0
qrcode>=8.2,<9.0
reportlab>=4.5,<5.0
pillow>=12.2,<13.0
